src/app/interceptor/jwt.interceptor.ts
import {Injectable} from '@angular/core';
import {HttpRequest, HttpHandler, HttpEvent, HttpInterceptor} from '@angular/common/http';
import {Observable} from 'rxjs';
import {AuthenticationService} from '../service/authentication.service';

@Injectable()
export class JwtInterceptor implements HttpInterceptor {

  constructor(private authenticationService: AuthenticationService) {
  }

  intercept(request: HttpRequest<any>, next: HttpHandler): Observable<HttpEvent<any>> {
    if (this.authenticationService.isLoggedIn) {
      request = request.clone({
        setHeaders: {
          Authorization: `Bearer ${localStorage.getItem('ACCESS_TOKEN')}`
        }
      });
    }

    return next.handle(request);
  }
}
